refactored to only have logical or logic and renamed. Dont need the exclusive or logic.
