+++ abis_mapping/plugins/__init__.py
@@ -5,7 +5,7 @@ from . import coordinates
 from . import chronological
 from . import empty
 from . import list
-from . import mutual_exclusion
+from . import logical_or
 from . import mutual_inclusion
 from . import tabular
 from . import timestamp
