+++ tests/plugins/test_logical_or.py
@@ -3,36 +3,26 @@
 
 # Third-Party
 import frictionless
-import pytest
 
 # Local
 from abis_mapping import plugins
 
 
-@pytest.mark.parametrize(
-    "use_or_behavior,n_invalid",
-    [
-        (False, 5),
-        (True, 1)
-    ]
-)
-def test_check_mutually_exclusive(use_or_behavior: bool, n_invalid: int) -> None:
-    """Tests the mutual exclusion checker."""
+def test_check_logical_or() -> None:
+    """Tests the logical or checker."""
     # Construct fake resource
     resource = frictionless.Resource(
         source=[
-            # Valid both
+            # Valid
             {"a": "A", "b": None, "c": None, "d": "D"},
             {"a": None, "b": "B", "c": None, "d": "D"},
             {"a": None, "b": None, "c": "C", "d": "D"},
-
-            # Invalid - not OR behavior
             {"a": "A", "b": "B", "c": None, "d": "D"},
             {"a": "A", "b": None, "c": "C", "d": "D"},
             {"a": None, "b": "B", "c": "C", "d": "D"},
             {"a": "A", "b": "B", "c": "C", "d": "D"},
 
-            # Invalid both
+            # Invalid
             {"a": None, "b": None, "c": None, "d": "D"},
         ]
     )
@@ -41,9 +31,8 @@ def test_check_mutually_exclusive(use_or_behavior: bool, n_invalid: int) -> None
     report = resource.validate(
         checklist=frictionless.Checklist(
             checks=[
-                plugins.mutual_exclusion.MutuallyExclusive(
+                plugins.logical_or.LogicalOr(
                     field_names=["a", "b", "c"],
-                    use_or_behavior=use_or_behavior,
                 )
             ]
         )
@@ -51,4 +40,4 @@ def test_check_mutually_exclusive(use_or_behavior: bool, n_invalid: int) -> None
 
     # Check
     assert not report.valid
-    assert len(report.flatten()) == n_invalid
+    assert len(report.flatten()) == 1
